feat: switch to rollup build with code-splitting
