+++ package.json
@@ -7,7 +7,7 @@
   "license": "MIT",
   "source": "src/index.tsx",
   "main": "dist/index.js",
-  "module": "dist/index.m.js",
+  "module": "dist/index.mjs",
   "types": "dist/index.d.ts",
   "keywords": [
     "react",
@@ -27,8 +27,10 @@
     "dist"
   ],
   "scripts": {
-    "start": "microbundle watch --raw --no-sourcemap --strict --no-compress --jsx React.createElement --format=cjs,esm",
-    "build": "microbundle --strict --jsx React.createElement --format=cjs,esm src/index.tsx",
+    "start": "rollup --config --watch",
+    "build": "rollup --config",
+    "start_old": "microbundle watch --raw --no-sourcemap --strict --no-compress --jsx React.createElement --format=cjs,esm",
+    "build_old": "microbundle --strict --jsx React.createElement --format=cjs,esm src/index.tsx",
     "test": "jest --passWithNoTests",
     "e2e": "playwright test",
     "lint": "eslint src test e2e --ext .ts,.tsx",
@@ -38,6 +40,7 @@
     "semantic-release": "semantic-release"
   },
   "dependencies": {
+    "@babel/runtime": "^7.17.2",
     "@virtuoso.dev/react-urx": "^0.2.12",
     "@virtuoso.dev/urx": "^0.2.12"
   },
@@ -45,11 +48,16 @@
     "react": ">=16"
   },
   "devDependencies": {
+    "@babel/core": "^7.17.5",
+    "@babel/plugin-transform-runtime": "^7.17.0",
+    "@babel/preset-env": "^7.16.11",
     "@commitlint/cli": "^12.1.4",
     "@commitlint/config-conventional": "^12.1.4",
     "@emotion/core": "^10.1.0",
     "@emotion/styled": "^10.0.27",
     "@playwright/test": "^1.17.1",
+    "@rollup/plugin-babel": "^5.3.0",
+    "@rollup/plugin-typescript": "^8.3.0",
     "@types/faker": "^5.1.4",
     "@types/jest": "^26.0.20",
     "@types/jsdom": "^16.2.3",
@@ -90,10 +98,12 @@
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
     "react-test-renderer": "^16.13.1",
+    "rollup": "^2.67.2",
+    "rollup-plugin-terser": "^7.0.2",
     "semantic-release": "^17.3.0",
     "ts-jest": "^26.5.1",
     "tslib": "^2.0.0",
-    "typescript": "^4.4.2"
+    "typescript": "~4.4.4"
   },
   "resolutions": {
     "@types/react": "^16.14.2"
