+++ src/components/CompProps.ts
@@ -1,0 +1,3 @@
+import { ForwardRefExoticComponent } from 'react'
+
+export type CompProps<T> = T extends ForwardRefExoticComponent<infer R> ? R : never
