+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export { Virtuoso } from './Virtuoso'
+export { GroupedVirtuoso } from './GroupedVirtuoso'
+export { TableVirtuoso } from './TableVirtuoso'
+export { VirtuosoGrid } from './VirtuosoGrid'
