+++ src/index.tsx
@@ -1,3 +1,3 @@
-export * from './components'
+export * from './components/index'
 export * from './interfaces'
 export { LogLevel } from './loggerSystem'
